app-web/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
Authlib==1.4.0
babel==2.16.0
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
email-validator
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.0
jsonschema_rs==0.22.3
MarkupSafe==3.0.2
motor==3.6.0
openai==1.55.1
orjson==3.10.12
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pymongo==4.9.2
python-multipart==0.0.19
redis
sniffio==1.3.1
starlette==0.39.2
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.30.6
websockets==13.1
authlib
motor
bcrypt
python-multipart
starlette-csrf
